Fix the billion laughs DoS attack, for Entity expansion.
This was already fixed in the REXML library of Ruby, but not in the native libxml library that is used by Nokogiri. Therefore parse the message first with the REXML in the Ruby library, by using Hash.from_xml. This will raise an “entity expansion has grown too large” RuntimeError.